+++ src/main/java/pigcart/puddleflood/config/ConfigData.java
@@ -1,7 +1,7 @@
 package pigcart.puddleflood.config;
 
-import pigcart.puddleflood.config.ConfigManager.ClientHasAuthority;
-import pigcart.puddleflood.config.ConfigManager.ReloadChunks;
+import pigcart.puddleflood.config.gui.ConfigResponders.ClientHasAuthority;
+import pigcart.puddleflood.config.gui.ConfigResponders.ReloadChunks;
 import pigcart.puddleflood.config.gui.Annotations.*;
 
 
@@ -15,6 +15,7 @@ public class ConfigData {
     public boolean doRainPuddles = true;
     public boolean doDripPuddles = true;
     public boolean doPotionPuddles = true;
+    @OnlyEditableIf(ClientHasAuthority.class)
     public boolean doCauldronPuddles = true;
     @OnlyEditableIf(ClientHasAuthority.class)
     public boolean doFlowingWaterPuddles = true;