+++ src/main/java/pigcart/puddleflood/duck/BlockTickDuck.java
@@ -1,8 +1,8 @@
 package pigcart.puddleflood.duck;
 
-import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
 
 public interface BlockTickDuck {
-    void puddleflood_clientRandomTick(ClientLevel level, BlockPos pos);
+    void puddleflood_clientRandomTick(Level level, BlockPos pos);
 }