+++ src/main/java/pigcart/puddleflood/loaders/forge/ForgeEntrypoint.java
@@ -1,33 +1,26 @@
 //? if forge {
 /*package pigcart.puddleflood.loaders.forge;
 
-import net.minecraft.client.Minecraft;
-import net.minecraft.client.renderer.ItemBlockRenderTypes;
-import net.minecraft.client.renderer.RenderType;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.CreativeModeTabs;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.client.ConfigScreenHandler;
-import net.minecraftforge.client.event.RegisterClientCommandsEvent;
-import net.minecraftforge.client.event.RegisterColorHandlersEvent;
 import net.minecraftforge.common.MinecraftForge;
-import net.minecraftforge.event.BuildCreativeModeTabContentsEvent;
-import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
-import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
+import net.minecraftforge.fml.loading.FMLEnvironment;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegisterEvent;
-import pigcart.puddleflood.PuddleFlood;
 import pigcart.puddleflood.block.PuddleBlock;
-import pigcart.puddleflood.config.ConfigManager;
+import pigcart.puddleflood.config.gui.ConfigScreen;
 
-import static pigcart.puddleflood.PuddleFlood.*;
+import static pigcart.puddleflood.PuddleFlood.MOD_ID;
+import static pigcart.puddleflood.PuddleFlood.PUDDLE_BLOCK;
 
 @Mod(MOD_ID)
 public class ForgeEntrypoint {
@@ -35,30 +28,23 @@ public class ForgeEntrypoint {
     public ForgeEntrypoint() {
         final IEventBus eventBus = FMLJavaModLoadingContext.get().getModEventBus();
 
-        // @SubscribeEvent is supposed to do this but it doesnt work
-        // idk what the deal is with the two different instances of IEventBus used here
-        MinecraftForge.EVENT_BUS.addListener(ForgeEntrypoint::onTickEvent);
-        MinecraftForge.EVENT_BUS.addListener(ForgeEntrypoint::onRegisterCommandsEvent);
         eventBus.addListener(ForgeEntrypoint::onRegisterEvent);
-        eventBus.addListener(ForgeEntrypoint::onBuildCreativeTabEvent);
-        eventBus.addListener(ForgeEntrypoint::onRegisterBlockColorEvent);
-        eventBus.addListener(ForgeEntrypoint::onRegisterItemColorEvent);
-        eventBus.addListener(ForgeEntrypoint::onSetupEvent);
 
-        ModLoadingContext.get().registerExtensionPoint(
-                ConfigScreenHandler.ConfigScreenFactory.class,
-                () -> new ConfigScreenHandler.ConfigScreenFactory(
-                        (client, parent) -> ConfigManager.screenPlease(parent)
-                )
-        );
-    }
+        if (FMLEnvironment.dist == Dist.CLIENT) {
+            MinecraftForge.EVENT_BUS.addListener(ForgeClientEntrypoint::onTickEvent);
+            MinecraftForge.EVENT_BUS.addListener(ForgeClientEntrypoint::onRegisterCommandsEvent);
+            eventBus.addListener(ForgeClientEntrypoint::onBuildCreativeTabEvent);
+            eventBus.addListener(ForgeClientEntrypoint::onRegisterBlockColorEvent);
+            eventBus.addListener(ForgeClientEntrypoint::onRegisterItemColorEvent);
+            eventBus.addListener(ForgeClientEntrypoint::onSetupEvent);
 
-    public static void onTickEvent(TickEvent.ClientTickEvent event) {
-        PuddleFlood.onTick(Minecraft.getInstance());
-    }
-
-    public static void onRegisterCommandsEvent(RegisterClientCommandsEvent event) {
-        event.getDispatcher().register(getCommands());
+            ModLoadingContext.get().registerExtensionPoint(
+                    ConfigScreenHandler.ConfigScreenFactory.class,
+                    () -> new ConfigScreenHandler.ConfigScreenFactory(
+                            (client, parent) -> ConfigScreen.get(parent)
+                    )
+            );
+        }
     }
 
     public static void onRegisterEvent(RegisterEvent event) {
@@ -74,24 +60,6 @@ public class ForgeEntrypoint {
                     h.register(new ResourceLocation(MOD_ID, "puddle"), blockItem);
                 }
         );
-    }
-
-    public static void onBuildCreativeTabEvent(BuildCreativeModeTabContentsEvent event) {
-        if (event.getTabKey() == CreativeModeTabs.NATURAL_BLOCKS) {
-            event.accept(PUDDLE_BLOCK);
-        }
-    }
-
-    public static void onRegisterBlockColorEvent(RegisterColorHandlersEvent.Block event){
-        event.register(PuddleFlood::puddleTintProvider, PUDDLE_BLOCK);
-    }
-
-    public static void onRegisterItemColorEvent(RegisterColorHandlersEvent.Item event){
-        event.register(PuddleFlood::puddleItemTintProvider, PUDDLE_BLOCK);
-    }
-
-    public static void onSetupEvent(FMLClientSetupEvent event) {
-        ItemBlockRenderTypes.setRenderLayer(PUDDLE_BLOCK, RenderType.translucent());
     }
 }
 *///?}