+++ src/main/java/pigcart/puddleflood/loaders/neoforge/NeoforgeEntrypoint.java
@@ -5,10 +5,12 @@ import net.minecraft.core.registries.Registries;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModLoadingContext;
 import net.neoforged.fml.common.Mod;
+import net.neoforged.fml.loading.FMLEnvironment;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 import net.neoforged.neoforge.registries.RegisterEvent;
+import pigcart.puddleflood.config.gui.ConfigScreen;
+import pigcart.puddleflood.loaders.fabric.FabricClientEntrypoint;
 import pigcart.puddleflood.loaders.fabric.FabricEntrypoint;
-import pigcart.puddleflood.config.ConfigManager;
 
 import static pigcart.puddleflood.PuddleFlood.MOD_ID;
 
@@ -18,7 +20,7 @@ public class NeoforgeEntrypoint {
     public NeoforgeEntrypoint(IEventBus eventBus) {
         ModLoadingContext.get().registerExtensionPoint(
                 IConfigScreenFactory.class,
-                () -> (modContainer, parent) -> ConfigManager.screenPlease(parent)
+                () -> (modContainer, parent) -> ConfigScreen.get(parent)
         );
         eventBus.addListener(RegisterEvent.class, NeoforgeEntrypoint::register);
     }
@@ -27,6 +29,9 @@ public class NeoforgeEntrypoint {
         // yeah blocks whatever ffapi has it covered
         event.register(Registries.BLOCK, h -> {
             FabricEntrypoint.init();
+            if (FMLEnvironment.dist.isClient()) {
+                FabricClientEntrypoint.init();
+            }
         });
     }
 }