+++ src/main/java/pigcart/puddleflood/mixin/client/BlockBehaviourMixin.java
@@ -1,7 +1,7 @@
 package pigcart.puddleflood.mixin.client;
 
-import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import org.spongepowered.asm.mixin.Mixin;
 import pigcart.puddleflood.duck.BlockTickDuck;
@@ -9,7 +9,7 @@ import pigcart.puddleflood.duck.BlockTickDuck;
 @Mixin(BlockBehaviour.class)
 public class BlockBehaviourMixin implements BlockTickDuck {
     @Override
-    public void puddleflood_clientRandomTick(ClientLevel level, BlockPos pos) {
+    public void puddleflood_clientRandomTick(Level level, BlockPos pos) {
 
     }
 }