+++ src/main/java/pigcart/puddleflood/mixin/client/BlockStateBaseMixin.java
@@ -1,7 +1,7 @@
 package pigcart.puddleflood.mixin.client;
 
-import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import org.spongepowered.asm.mixin.Mixin;
@@ -15,7 +15,7 @@ public abstract class BlockStateBaseMixin implements BlockTickDuck {
     public abstract Block getBlock();
 
     @Override
-    public void puddleflood_clientRandomTick(ClientLevel level, BlockPos pos) {
+    public void puddleflood_clientRandomTick(Level level, BlockPos pos) {
         ((BlockTickDuck)this.getBlock()).puddleflood_clientRandomTick(level, pos);
     }
 }