+++ src/main/java/pigcart/puddleflood/mixin/client/ClientPacketListenerMixin.java
@@ -8,7 +8,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import pigcart.puddleflood.PuddleFlood;
-import pigcart.puddleflood.config.ConfigManager;
 
 @Mixin(ClientPacketListener.class)
 public class ClientPacketListenerMixin {